+++ src/aui-base/js/aui-loader.js
@@ -384,7 +384,8 @@ Y.mix(YUI.Env[Y.version].modules, {
         "requires": [
             "aui-datatable-base-cell-editor",
             "escape"
-        ]
+        ],
+        "skinnable": true
     },
     "aui-datatable-body": {
         "requires": [
@@ -2130,4 +2131,4 @@ Y.mix(YUI.Env[Y.version].modules, {
         ]
     }
 });
-YUI.Env[Y.version].md5 = 'a088be86166e18d96a89b4e8a5199400';
+YUI.Env[Y.version].md5 = '0ebd18e5b0518b03ca8dcf7332a5901b';